requirements.txt
flask
langchain
google-generativeai
sentence-transformers
faiss-cpu
python-dotenv
PyPDF2
